Safari-611.1.10/LayoutTests/http/tests/security/contentSecurityPolicy/script-src-blocked-error-event-expected.txt

CONSOLE MESSAGE: Refused to load http://localhost:8000/security/contentSecurityPolicy/resources/script-set-value.js because it does not appear in the script-src directive of the Content Security Policy.

PASS Script element has error on bad script-src
